app/pyproject.toml
[project]
name = "ocr_compare"
version = "0.1.0"
description = "OCR Compare"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.16",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "docx2python>=3.5.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langid>=1.1.6",
    "matplotlib>=3.10.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pip>=25.0.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.3",
    "pyjwt>=2.10.1",
    "pytesseract>=0.3.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.1",
    "websockets>=15.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "session"
